Add documentation for the new sign in methods
