+++ app/src/androidTest/java/com/github/h3lp3rs/h3lp/signIn/SignInActivitySignedInUserTest.kt
@@ -1,22 +1,21 @@
-package com.github.h3lp3rs.h3lp
+package com.github.h3lp3rs.h3lp.signIn
 
-import android.app.Activity
 import android.content.Intent
-import androidx.activity.result.ActivityResult
+import androidx.test.core.app.ApplicationProvider
 import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.action.ViewActions.click
 import androidx.test.espresso.intent.Intents
-import androidx.test.espresso.intent.matcher.IntentMatchers.*
-import androidx.test.espresso.matcher.ViewMatchers.*
+import androidx.test.espresso.intent.matcher.IntentMatchers
+import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.github.h3lp3rs.h3lp.signIn.GoogleSignInAdaptor
+import com.github.h3lp3rs.h3lp.MainPageActivity
+import com.github.h3lp3rs.h3lp.R
 import com.github.h3lp3rs.h3lp.signIn.SignIn
 import com.github.h3lp3rs.h3lp.signIn.SignInActivity
 import com.github.h3lp3rs.h3lp.signIn.SignInInterface
 import com.google.android.gms.tasks.Task
 import com.google.firebase.auth.AuthResult
-import junit.framework.Assert.*
 import org.junit.After
 import org.junit.Before
 import org.junit.Rule
@@ -26,12 +25,8 @@ import org.mockito.Mockito
 import org.mockito.Mockito.`when`
 import org.mockito.kotlin.anyOrNull
 
-
 @RunWith(AndroidJUnit4::class)
-class SignInWithGoogleTest {
-    private lateinit var intent : Intent
-    private val googleSignInPackageName = "com.google.android.gms"
-    private var authenticationStarted = false
+class SignInActivitySignedInUserTest {
 
     @get:Rule
     val testRule = ActivityScenarioRule(
@@ -43,40 +38,24 @@ class SignInWithGoogleTest {
         Intents.init()
 
         val signInMock = Mockito.mock(SignInInterface::class.java)
-        `when`(signInMock.isSignedIn()).thenReturn(false)
+        `when`(signInMock.isSignedIn()).thenReturn(true)
 
         testRule.scenario.onActivity { activity ->
-            intent  = GoogleSignInAdaptor.signIn(activity)
-            val taskG = GoogleSignInAdaptor.authenticate(ActivityResult(Activity.RESULT_OK,intent),activity)
+            val intent  = Intent(ApplicationProvider.getApplicationContext(), activity.javaClass)
             val taskMock = Mockito.mock(Task::class.java)
             `when`(taskMock.isSuccessful).thenReturn(true)
             `when`(taskMock.isComplete).thenReturn(true)
             `when`(signInMock.signIn(activity)).thenReturn(intent)
-            `when`(signInMock.authenticate(anyOrNull(), anyOrNull())).thenAnswer{
-                authenticationStarted = true
-                taskMock
-            }
+            `when`(signInMock.authenticate(anyOrNull(), anyOrNull())).thenReturn(taskMock)
         }
 
         SignIn.set(signInMock as SignInInterface<AuthResult>)
     }
 
     @Test
-    fun signInWithGoogleLaunchesCorrectIntent(){
-        clickSignInButton()
-        Intents.intended(hasPackage(googleSignInPackageName))
-    }
-
-    @Test
-    fun signInWithGoogleLaunchesAuthenticationProcess(){
-        clickSignInButton()
-        assertNotNull(GoogleSignInAdaptor.gso)
-        testRule.scenario.onActivity { activity -> activity.authenticateUser(ActivityResult(Activity.RESULT_OK,intent),activity)}
-        assert(authenticationStarted)
-    }
-
-    private fun clickSignInButton(){
+    fun signedInUserMoveToMainPageDirectly(){
         onView(withId(R.id.signInButton)).perform(click())
+        Intents.intended(IntentMatchers.hasComponent(MainPageActivity::class.java.name))
     }
 
     @After
@@ -84,7 +63,3 @@ class SignInWithGoogleTest {
         Intents.release()
     }
 }
-
-
-
-
