+++ app/src/androidTest/java/com/github/h3lp3rs/h3lp/signIn/SignInWithGoogleTest.kt
@@ -1,21 +1,23 @@
-package com.github.h3lp3rs.h3lp
+package com.github.h3lp3rs.h3lp.signIn
 
 import android.app.Activity
 import android.content.Intent
 import androidx.activity.result.ActivityResult
-import androidx.test.core.app.ApplicationProvider
 import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.action.ViewActions.click
 import androidx.test.espresso.intent.Intents
-import androidx.test.espresso.intent.matcher.IntentMatchers
-import androidx.test.espresso.matcher.ViewMatchers.withId
+import androidx.test.espresso.intent.matcher.IntentMatchers.*
+import androidx.test.espresso.matcher.ViewMatchers.*
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import com.github.h3lp3rs.h3lp.R
+import com.github.h3lp3rs.h3lp.signIn.GoogleSignInAdaptor
 import com.github.h3lp3rs.h3lp.signIn.SignIn
 import com.github.h3lp3rs.h3lp.signIn.SignInActivity
 import com.github.h3lp3rs.h3lp.signIn.SignInInterface
 import com.google.android.gms.tasks.Task
 import com.google.firebase.auth.AuthResult
+import junit.framework.Assert.*
 import org.junit.After
 import org.junit.Before
 import org.junit.Rule
@@ -25,10 +27,11 @@ import org.mockito.Mockito
 import org.mockito.Mockito.`when`
 import org.mockito.kotlin.anyOrNull
 
-@RunWith(AndroidJUnit4::class)
-class SignInActivityNewUserTest {
 
-    private lateinit var intent: Intent
+@RunWith(AndroidJUnit4::class)
+class SignInWithGoogleTest {
+    private lateinit var intent : Intent
+    private val googleSignInPackageName = "com.google.android.gms"
     private var authenticationStarted = false
 
     @get:Rule
@@ -44,7 +47,8 @@ class SignInActivityNewUserTest {
         `when`(signInMock.isSignedIn()).thenReturn(false)
 
         testRule.scenario.onActivity { activity ->
-            intent  = Intent(ApplicationProvider.getApplicationContext(), activity.javaClass)
+            intent  = GoogleSignInAdaptor.signIn(activity)
+            val taskG = GoogleSignInAdaptor.authenticate(ActivityResult(Activity.RESULT_OK,intent),activity)
             val taskMock = Mockito.mock(Task::class.java)
             `when`(taskMock.isSuccessful).thenReturn(true)
             `when`(taskMock.isComplete).thenReturn(true)
@@ -59,18 +63,15 @@ class SignInActivityNewUserTest {
     }
 
     @Test
-    fun newUserSignInLaunchesCorrectIntent(){
+    fun signInWithGoogleLaunchesCorrectIntent(){
         clickSignInButton()
-        Intents.intended(IntentMatchers.hasComponent(SignInActivity::class.java.name))
+        Intents.intended(hasPackage(googleSignInPackageName))
     }
 
     @Test
-    fun newUserSignInLaunchesAuthenticationProcess(){
-        /*val resultData = Intent()
-        val result = Instrumentation.ActivityResult(Activity.RESULT_OK, resultData)
-        intending(anyIntent()).respondWith(result)
-*/
+    fun signInWithGoogleLaunchesAuthenticationProcess(){
         clickSignInButton()
+        assertNotNull(GoogleSignInAdaptor.gso)
         testRule.scenario.onActivity { activity -> activity.authenticateUser(ActivityResult(Activity.RESULT_OK,intent),activity)}
         assert(authenticationStarted)
     }
@@ -84,3 +85,7 @@ class SignInActivityNewUserTest {
         Intents.release()
     }
 }
+
+
+
+
